Move table_name_to_class_name to utils and add tests
